+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Fixed
+
+- `PlumeKeyProvider` was providing duplicates - this is now fixed.
+
 ### Changed
 
 - Duplicates are now handled the same in `TigerGraphDriver` as the rest